remove_stupid_pointless_stuff
